create deck

+++ src/main/java/com/example/produktmicroservice/repository/PokemonCardRepository.java
@@ -3,5 +3,5 @@ package com.example.produktmicroservice.repository;
 import com.example.produktmicroservice.entity.PokemonCard;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PokemonCardRepository extends JpaRepository<PokemonCard,Integer> {
+public interface PokemonCardRepository extends JpaRepository<PokemonCard,Long> {
 }
