+++ src/main/java/com/example/produktmicroservice/repository/PokemonDeckRepository.java
@@ -3,5 +3,5 @@ package com.example.produktmicroservice.repository;
 import com.example.produktmicroservice.entity.PokemonDeck;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PokemonDeckRepository extends JpaRepository<PokemonDeck,Integer> {
+public interface PokemonDeckRepository extends JpaRepository<PokemonDeck,Long> {
 }
